requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
astroid==3.1.0
attrs==23.2.0
autoflake==2.3.0
black==24.2.0
blinker==1.7.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
dataclasses-json==0.6.4
dill==0.3.8
distro==1.9.0
docutils==0.20.1
filelock==3.13.1
flake8==7.0.0
Flask==3.0.2
Flask-DebugToolbar==0.14.1
Flask-Testing==0.8.1
frozenlist==1.4.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib_metadata==7.0.2
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
jaraco.classes==3.3.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
keyring==24.3.1
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.42
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langcodes==3.3.0
langsmith==0.1.23
lxml==4.9.4
macholib==1.16.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nh3==0.2.15
numpy==1.26.4
openai==1.13.3
orjson==3.9.15
packaging==23.2
pathspec==0.12.1
phonenumbers==8.13.31
pkginfo==1.10.0
platformdirs==4.2.0
pluggy==1.4.0
preshed==3.0.9
presidio-analyzer==2.2.353
presidio-anonymizer==2.2.353
pycodestyle==2.11.1
pycryptodome==3.20.0
pydantic==1.10.14
pydantic_core==2.16.3
pyflakes==3.2.0
Pygments==2.17.2
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.5
pylint==3.1.0
pytest==8.1.1
python-dotenv==1.0.1
PyYAML==6.0.1
readme_renderer==43.0
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
setuptools==69.1.1
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.28
srsly==2.4.8
tenacity==8.2.3
thinc==8.2.3
tiktoken==0.5.2
tldextract==5.1.1
tomlkit==0.12.4
tqdm==4.66.2
twine==5.0.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
